refactor: 중복 클래스 제거
